Archiver config files moved to folder and some classes moved to other package
