--- src/test/resources/archiver-test-debug-mode.failconf
@@ -1,37 +1,0 @@
-##### Image Datastore Configuration #####
-datastore_url_prefix=
-datastore_ip=
-datastore_port=
-datastore_name=
-datastore_driver=
-datastore_username=
-datastore_password=
-
-saps_execution_period_archiver=
-saps_execution_period_garbage_collector=
-saps_execution_debug_mode=true
-
-saps_permanent_storage_type=swift
-
-# NFS
-saps_export_path=src/test/resources/archiver-test
-
-# Configuração do sft
-swift_folder_prefix=archiver
-swift_container_name=saps-test
-swift_input_pseud_folder_prefix=
-swift_output_pseud_folder_prefix=
-swift_username=
-swift_password=
-swift_tenant_id=
-swift_tenant_name=
-swift_auth_url=
-
-# Configuração do fogbow keystone
-fogbow.keystonev3.project.id=
-fogbow.keystonev3.user.id=
-fogbow.keystonev3.password=
-fogbow.keystonev3.auth.url=
-fogbow.keystonev3.swift.url=
-fogbow_cli_path=
-fogbow.keystonev3.swift.token.update.period=
